Move layout out of PageHome
