+++ src/main/javascript/src/components/ui/AppLayout/app-layout.scss
@@ -1,4 +1,4 @@
-.home_page-home {
+.ui_app-layout {
   &__footer {
     text-align: center;
   }
