+++ src/main/javascript/src/index.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.scss';
-import { Routes } from './components/Routes';
+import { Routes } from './Routes';
 
 ReactDOM.render(<Routes />, document.getElementById('root'));
